+++ src/components/header/header.module.scss
@@ -15,7 +15,7 @@
   font-size: 35px;
   font-weight: bold;
   color: $white;
-  border-bottom: 4px solid purple;
+  // border-bottom: 4px solid purple;
 }
 
 .title:active {
