+++ src/components/modals/signOutModal/signOutModal.module.scss
@@ -2,7 +2,6 @@
 
 .container {
   margin: 0 auto;
-  padding: 25px;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -13,6 +12,7 @@
 }
 
 .title {
+  padding: 0 15px;
   font-size: 20px;
   font-weight: bold;
 }
