+++ src/styles/formStyles.scss
@@ -2,8 +2,8 @@
 
 .form {
   width: 100%;
-  padding: 40px;
   display: flex;
+  padding: 0 0 40px 0;
   flex-direction: column;
   align-items: center;
   box-shadow: 0 0 15px 10px #562370;
@@ -12,7 +12,7 @@
 }
 
 .formTitle {
-  align-self: flex-start;
+  align-self: center;
   padding-bottom: 80px;
   color: $white;
   font-size: 40px;
