+++ src/styles/main.scss
@@ -30,6 +30,6 @@ h2 {
   flex-direction: column;
 }
 
-.active {
-  border-bottom: 4px solid purple;
-}
+// .active {
+//   border-bottom: 4px solid purple;
+// }
